feat: :sparkles: rewrite prebuild script into cron job to update supabase db
